+++ src/Opulence/Framework/Console/Bootstrappers/CommandsBootstrapper.php
@@ -18,6 +18,7 @@ use Opulence\Framework\Console\Commands\AppUpCommand;
 use Opulence\Framework\Console\Commands\ComposerDumpAutoloadCommand;
 use Opulence\Framework\Console\Commands\ComposerUpdateCommand;
 use Opulence\Framework\Console\Commands\EncryptionKeyGenerationCommand;
+use Opulence\Framework\Console\Commands\EncryptionPasswordGenerationCommand;
 use Opulence\Framework\Console\Commands\FlushFrameworkCacheCommand;
 use Opulence\Framework\Console\Commands\FlushViewCacheCommand;
 use Opulence\Framework\Console\Commands\MakeCommandCommand;
@@ -42,6 +43,7 @@ class CommandsBootstrapper extends Bootstrapper
         ComposerDumpAutoloadCommand::class,
         ComposerUpdateCommand::class,
         EncryptionKeyGenerationCommand::class,
+        EncryptionPasswordGenerationCommand::class,
         FlushFrameworkCacheCommand::class,
         FlushViewCacheCommand::class,
         MakeCommandCommand::class,
@@ -50,7 +52,7 @@ class CommandsBootstrapper extends Bootstrapper
         MakeEntityCommand::class,
         MakeHttpMiddlewareCommand::class,
         RenameAppCommand::class,
-        UUidGenerationCommand::class
+        UuidGenerationCommand::class
     ];
     /** @var CommandCollection The list of console commands */
     private $commandCollection = null;