+++ src/Opulence/Framework/Console/Commands/EncryptionKeyGenerationCommand.php
@@ -16,22 +16,11 @@ use Opulence\Console\Responses\IResponse;
 
 /**
  * Defines the encryption key generator command
+ * 
+ * @deprecated since v1.0.0-beta4
  */
-class EncryptionKeyGenerationCommand extends Command
+class EncryptionKeyGenerationCommand extends EncryptionPasswordGenerationCommand
 {
-    /** @var Paths The application paths */
-    private $paths = null;
-
-    /**
-     * @param Paths $paths The application paths
-     */
-    public function __construct(Paths $paths)
-    {
-        parent::__construct();
-
-        $this->paths = $paths;
-    }
-
     /**
      * @inheritdoc
      */
@@ -52,16 +41,10 @@ class EncryptionKeyGenerationCommand extends Command
      */
     protected function doExecute(IResponse $response)
     {
-        $key = bin2hex(random_bytes(16));
-        $environmentConfigPath = $this->paths["config"] . "/environment/.env.app.php";
-
-        if (!$this->optionIsSet("show") && file_exists($environmentConfigPath)) {
-            $contents = file_get_contents($environmentConfigPath);
-            $newContents = preg_replace("/\"ENCRYPTION_KEY\",\s*\"[^\"]*\"/U", '"ENCRYPTION_KEY", "' . $key . '"',
-                $contents);
-            file_put_contents($environmentConfigPath, $newContents);
-        }
-
-        $response->writeln("Generated key: <info>$key</info>");
+        $deprecationMessage = "EncryptionKeyGenerationCommand has been deprecated.  " .
+            "Use EncryptionPasswordGenerationCommand instead.";
+        trigger_error($deprecationMessage, E_USER_DEPRECATED);
+        parent::doExecute($response);
+        $response->writeln("<warning>$deprecationMessage</warning>");
     }
 }